Removed floating point math from calculation

+++ src/statsdef.h
@@ -338,7 +338,7 @@ struct PROPULSION_STATS : public COMPONENT_STATS
 	unsigned skidDeceleration = 0;
 	unsigned deceleration = 0;
 	unsigned acceleration = 0;
-	float repairFactor = 0;
+	unsigned repairFactor = 0;
 
 	struct UPGRADE : COMPONENT_STATS::UPGRADE
 	{
